requirements.txt
pandas
google-api-python-client
google-genai
chromadb
tqdm
python-dotenv
streamlit